requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
Werkzeug==2.3.7
requests==2.31.0
google-generativeai==0.3.1
markdown==3.5.1
pytz==2023.3.post1
cryptography==41.0.5
python-dotenv==1.0.0
APScheduler==3.10.4
tweepy==4.15.0
beautifulsoup4==4.13.3

# Database and Caching
redis>=5.0.1
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0
